oer-core/src/main/java/org/emschu/oer/core/util/Hasher.java
package org.emschu.oer.core.util;

/*-
 * #%L
 * oer-server
 * %%
 * Copyright (C) 2019 emschu[aet]mailbox.org
 * %%
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 * 
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 * 
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 * #L%
 */

import javax.xml.bind.DatatypeConverter;
import java.security.MessageDigest;
import java.security.NoSuchAlgorithmException;
import java.util.logging.Logger;

/**
 * this class is used to trivially create an md5 hash of a string
 */
public class Hasher {
    private static final Logger LOG = Logger.getLogger(Hasher.class.getName());

    /**
     * empty string = null hash
     *
     * @param text
     * @return
     */
    public static final String getHash(String text) {
        if (text == null) {
            text = "";
        }
        MessageDigest md = null;
        try {
            md = MessageDigest.getInstance("MD5");
            md.update(text.getBytes());
            byte[] digest = md.digest();
            return DatatypeConverter.printHexBinary(digest).toUpperCase();
        } catch (NoSuchAlgorithmException e) {
            LOG.warning("MD5 algorithm not found.");
        }
        throw new IllegalStateException(String.format("could not create hash for text '%s'", text));
    }
}
